Move gnprintf/atcprintf from core to logic.
