+++ core/logic/AMBuilder
@@ -75,6 +75,7 @@ binary.sources += [
   'Logger.cpp',
   'smn_core.cpp',
   'smn_menus.cpp',
+  'sprintf.cpp',
 ]
 if builder.target_platform == 'windows':
   binary.sources += ['thread/WinThreads.cpp']
